desktop-app/requirements.txt
PyQt6>=6.4.0
python-dotenv>=0.19.0
requests>=2.26.0
flask==3.0.0
python-dateutil==2.8.2
customtkinter==5.2.2
pillow==10.2.0 